test(first-party): address gsuite fixture review
