+++ crates/ironclaw_first_party_extensions/tests/gsuite_core.rs
@@ -1,13 +1,11 @@
-use std::{
-    collections::VecDeque,
-    sync::{Arc, Mutex},
-};
+mod support;
+
+use std::sync::Arc;
 
 use ironclaw_auth::{
-    AuthProductScope, AuthProviderId, AuthSurface, CredentialAccountLabel,
-    CredentialAccountService, CredentialAccountStatus, CredentialOwnership,
-    GOOGLE_CALENDAR_READONLY_SCOPE, GOOGLE_GMAIL_MODIFY_SCOPE, GOOGLE_GMAIL_READONLY_SCOPE,
-    GOOGLE_GMAIL_SEND_SCOPE, InMemoryAuthProductServices, NewCredentialAccount, ProviderScope,
+    AuthProviderId, CredentialAccountStatus, GOOGLE_CALENDAR_READONLY_SCOPE,
+    GOOGLE_GMAIL_MODIFY_SCOPE, GOOGLE_GMAIL_READONLY_SCOPE, GOOGLE_GMAIL_SEND_SCOPE,
+    InMemoryAuthProductServices,
 };
 use ironclaw_first_party_extensions::{
     CALENDAR_ADD_ATTENDEES_CAPABILITY_ID, CALENDAR_CREATE_EVENT_CAPABILITY_ID,
@@ -18,236 +16,14 @@ use ironclaw_first_party_extensions::{
     GMAIL_GET_MESSAGE_CAPABILITY_ID, GMAIL_LIST_MESSAGES_CAPABILITY_ID,
     GMAIL_REPLY_TO_MESSAGE_CAPABILITY_ID, GMAIL_SEND_MESSAGE_CAPABILITY_ID,
     GMAIL_TRASH_MESSAGE_CAPABILITY_ID, GSUITE_OUTPUT_BYTES_LIMIT, GSUITE_RESPONSE_BODY_LIMIT,
-    GsuiteDispatchError, GsuiteDispatchRequest, GsuiteExecutor, google_provider_id,
-    gsuite_package_specs, gsuite_resource_profile,
+    GsuiteDispatchRequest, GsuiteExecutor, gsuite_package_specs, gsuite_resource_profile,
 };
 use ironclaw_host_api::{
-    CapabilityId, InvocationId, NetworkMethod, NetworkScheme,
-    RUNTIME_HTTP_REASON_RESPONSE_BODY_LIMIT_EXCEEDED, ResourceScope, RuntimeDispatchErrorKind,
-    RuntimeHttpEgress, RuntimeHttpEgressError, RuntimeHttpEgressRequest, RuntimeHttpEgressResponse,
-    SecretHandle, UserId,
+    NetworkMethod, NetworkScheme, RUNTIME_HTTP_REASON_RESPONSE_BODY_LIMIT_EXCEEDED,
+    RuntimeDispatchErrorKind, RuntimeHttpEgressError, SecretHandle,
 };
-use serde_json::{Value, json};
-
-#[derive(Default)]
-struct RecordingEgress {
-    requests: Mutex<Vec<RuntimeHttpEgressRequest>>,
-    responses: Mutex<VecDeque<Result<RuntimeHttpEgressResponse, RuntimeHttpEgressError>>>,
-}
-
-impl RecordingEgress {
-    fn with_responses(responses: Vec<RuntimeHttpEgressResponse>) -> Self {
-        Self {
-            requests: Mutex::new(Vec::new()),
-            responses: Mutex::new(responses.into_iter().map(Ok).collect()),
-        }
-    }
-
-    fn with_errors(errors: Vec<RuntimeHttpEgressError>) -> Self {
-        Self {
-            requests: Mutex::new(Vec::new()),
-            responses: Mutex::new(errors.into_iter().map(Err).collect()),
-        }
-    }
-
-    fn json(body: serde_json::Value) -> RuntimeHttpEgressResponse {
-        Self::json_with_request_bytes(body, 123)
-    }
-
-    fn json_status(status: u16, body: serde_json::Value) -> RuntimeHttpEgressResponse {
-        let body = serde_json::to_vec(&body).expect("response body serializes");
-        RuntimeHttpEgressResponse {
-            status,
-            headers: Vec::new(),
-            request_bytes: 123,
-            response_bytes: body.len() as u64,
-            body,
-            redaction_applied: true,
-        }
-    }
-
-    fn empty(status: u16) -> RuntimeHttpEgressResponse {
-        RuntimeHttpEgressResponse {
-            status,
-            headers: Vec::new(),
-            body: Vec::new(),
-            request_bytes: 123,
-            response_bytes: 0,
-            redaction_applied: true,
-        }
-    }
-
-    fn json_with_request_bytes(
-        body: serde_json::Value,
-        request_bytes: u64,
-    ) -> RuntimeHttpEgressResponse {
-        let body = serde_json::to_vec(&body).expect("response body serializes");
-        RuntimeHttpEgressResponse {
-            status: 200,
-            headers: Vec::new(),
-            request_bytes,
-            response_bytes: body.len() as u64,
-            body,
-            redaction_applied: true,
-        }
-    }
-
-    fn malformed_json() -> RuntimeHttpEgressResponse {
-        RuntimeHttpEgressResponse {
-            status: 200,
-            headers: Vec::new(),
-            request_bytes: 123,
-            response_bytes: 1,
-            body: b"{".to_vec(),
-            redaction_applied: true,
-        }
-    }
-
-    fn requests(&self) -> Vec<RuntimeHttpEgressRequest> {
-        self.requests.lock().expect("egress lock").clone()
-    }
-}
-
-impl RuntimeHttpEgress for RecordingEgress {
-    fn execute(
-        &self,
-        request: RuntimeHttpEgressRequest,
-    ) -> Result<RuntimeHttpEgressResponse, RuntimeHttpEgressError> {
-        self.requests.lock().expect("egress lock").push(request);
-        self.responses
-            .lock()
-            .expect("response lock")
-            .pop_front()
-            .unwrap_or_else(|| Ok(RecordingEgress::json(json!({"id":"sent-message"}))))
-    }
-}
-
-fn scope() -> ResourceScope {
-    ResourceScope::local_default(UserId::new("alice").unwrap(), InvocationId::new()).unwrap()
-}
-
-fn auth_scope(scope: &ResourceScope) -> AuthProductScope {
-    AuthProductScope::new(scope.clone(), AuthSurface::Api)
-}
-
-fn provider_scope(value: &str) -> ProviderScope {
-    ProviderScope::new(value).unwrap()
-}
-
-fn capability_id(value: &str) -> CapabilityId {
-    CapabilityId::new(value).unwrap()
-}
-
-fn fixture(area: &str, name: &str) -> Value {
-    let path = format!(
-        "{}/tests/fixtures/google_api/{area}/{name}",
-        env!("CARGO_MANIFEST_DIR")
-    );
-    let raw = std::fs::read_to_string(&path).unwrap_or_else(|e| panic!("read fixture {path}: {e}"));
-    serde_json::from_str(&raw).unwrap_or_else(|e| panic!("parse fixture {path}: {e}"))
-}
-
-async fn auth_with_google_account(
-    scope: &ResourceScope,
-    scopes: Vec<ProviderScope>,
-) -> Arc<InMemoryAuthProductServices> {
-    let auth = Arc::new(InMemoryAuthProductServices::new());
-    add_google_account(
-        &auth,
-        scope,
-        "work google",
-        scopes,
-        CredentialAccountStatus::Configured,
-        true,
-    )
-    .await;
-    auth
-}
-
-async fn auth_with_google_account_status(
-    scope: &ResourceScope,
-    scopes: Vec<ProviderScope>,
-    status: CredentialAccountStatus,
-    include_access_secret: bool,
-) -> Arc<InMemoryAuthProductServices> {
-    let auth = Arc::new(InMemoryAuthProductServices::new());
-    add_google_account(
-        &auth,
-        scope,
-        "work google",
-        scopes,
-        status,
-        include_access_secret,
-    )
-    .await;
-    auth
-}
-
-async fn add_google_account(
-    auth: &InMemoryAuthProductServices,
-    scope: &ResourceScope,
-    label: &str,
-    scopes: Vec<ProviderScope>,
-    status: CredentialAccountStatus,
-    include_access_secret: bool,
-) {
-    auth.create_account(NewCredentialAccount {
-        scope: auth_scope(scope),
-        provider: google_provider_id().unwrap(),
-        label: CredentialAccountLabel::new(label).unwrap(),
-        status,
-        ownership: CredentialOwnership::UserReusable,
-        owner_extension: None,
-        granted_extensions: Vec::new(),
-        access_secret: include_access_secret
-            .then(|| SecretHandle::new("google-access-token").unwrap()),
-        refresh_secret: None,
-        scopes,
-    })
-    .await
-    .unwrap();
-}
-
-async fn dispatch_ok(
-    auth: Arc<InMemoryAuthProductServices>,
-    scope: ResourceScope,
-    capability: &str,
-    input: serde_json::Value,
-    egress: Arc<RecordingEgress>,
-) -> serde_json::Value {
-    let executor = GsuiteExecutor::new(auth);
-    let capability_id = capability_id(capability);
-    executor
-        .dispatch(GsuiteDispatchRequest {
-            capability_id: &capability_id,
-            scope: &scope,
-            input: &input,
-            runtime_http_egress: egress,
-        })
-        .await
-        .unwrap()
-        .output
-}
-
-async fn dispatch_error(
-    auth: Arc<InMemoryAuthProductServices>,
-    scope: ResourceScope,
-    capability: &str,
-    input: serde_json::Value,
-    egress: Arc<RecordingEgress>,
-) -> GsuiteDispatchError {
-    let executor = GsuiteExecutor::new(auth);
-    let capability_id = capability_id(capability);
-    executor
-        .dispatch(GsuiteDispatchRequest {
-            capability_id: &capability_id,
-            scope: &scope,
-            input: &input,
-            runtime_http_egress: egress,
-        })
-        .await
-        .unwrap_err()
-}
+use serde_json::json;
+use support::*;
 
 #[test]
 fn gsuite_packages_declare_calendar_and_gmail_capabilities() {
@@ -955,267 +731,4 @@ async fn add_attendees_reports_both_google_api_requests() {
 
     assert_eq!(egress.requests().len(), 2);
     assert_eq!(result.usage.network_egress_bytes, 312);
-}
-
-#[tokio::test]
-async fn calendar_read_handlers_use_recorded_google_api_shapes() {
-    let scope = scope();
-    let auth =
-        auth_with_google_account(&scope, vec![provider_scope(GOOGLE_CALENDAR_READONLY_SCOPE)])
-            .await;
-    let egress = Arc::new(RecordingEgress::with_responses(vec![
-        RecordingEgress::json_status(200, fixture("calendar", "calendar_list.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "events_list.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "event_get.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "free_busy.json")),
-    ]));
-
-    let calendars = dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_LIST_CALENDARS_CAPABILITY_ID,
-        json!({}),
-        egress.clone(),
-    )
-    .await;
-    let events = dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_LIST_EVENTS_CAPABILITY_ID,
-        json!({
-            "calendar_id": "primary",
-            "time_min": "2026-05-21T00:00:00Z",
-            "time_max": "2026-05-22T00:00:00Z",
-            "max_results": 50
-        }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_GET_EVENT_CAPABILITY_ID,
-        json!({ "calendar_id": "primary", "event_id": "evt-standup-001" }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth,
-        scope,
-        CALENDAR_FIND_FREE_SLOTS_CAPABILITY_ID,
-        json!({
-            "timeMin": "2026-05-21T09:00:00Z",
-            "timeMax": "2026-05-21T17:00:00Z",
-            "items": [{ "id": "primary" }]
-        }),
-        egress.clone(),
-    )
-    .await;
-
-    assert_eq!(calendars["body"]["items"][0]["id"], "primary");
-    assert_eq!(events["body"]["nextPageToken"], "CiAKGjBpNDd2Nm");
-    let requests = egress.requests();
-    assert_eq!(requests.len(), 4);
-    assert_eq!(requests[0].method, NetworkMethod::Get);
-    assert!(requests[0].url.ends_with("/users/me/calendarList"));
-    assert!(requests[1].url.contains("/calendars/primary/events"));
-    assert!(requests[1].url.contains("timeMin=2026-05-21T00%3A00%3A00Z"));
-    assert!(requests[1].url.contains("maxResults=50"));
-    assert!(requests[2].url.ends_with("/events/evt-standup-001"));
-    assert_eq!(requests[3].method, NetworkMethod::Post);
-    assert!(requests[3].url.ends_with("/freeBusy"));
-}
-
-#[tokio::test]
-async fn calendar_write_handlers_use_recorded_google_api_shapes() {
-    let scope = scope();
-    let auth = auth_with_google_account(
-        &scope,
-        vec![provider_scope(ironclaw_auth::GOOGLE_CALENDAR_EVENTS_SCOPE)],
-    )
-    .await;
-    let egress = Arc::new(RecordingEgress::with_responses(vec![
-        RecordingEgress::json_status(200, fixture("calendar", "event_created.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "event_created.json")),
-        RecordingEgress::empty(204),
-        RecordingEgress::json_status(200, fixture("calendar", "event_get.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "event_created.json")),
-        RecordingEgress::json_status(200, fixture("calendar", "event_created.json")),
-    ]));
-
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_CREATE_EVENT_CAPABILITY_ID,
-        json!({ "calendar_id": "primary", "event": { "summary": "Project review" } }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_UPDATE_EVENT_CAPABILITY_ID,
-        json!({
-            "calendar_id": "primary",
-            "event_id": "evt-001",
-            "event": { "summary": "Updated review" }
-        }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_DELETE_EVENT_CAPABILITY_ID,
-        json!({ "calendar_id": "primary", "event_id": "evt-001" }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        CALENDAR_ADD_ATTENDEES_CAPABILITY_ID,
-        json!({
-            "calendar_id": "primary",
-            "event_id": "evt-001",
-            "attendees": [{ "email": "ada@example.com" }]
-        }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth,
-        scope,
-        CALENDAR_SET_REMINDER_CAPABILITY_ID,
-        json!({
-            "calendar_id": "primary",
-            "event_id": "evt-001",
-            "reminders": {
-                "useDefault": false,
-                "overrides": [{ "method": "popup", "minutes": 10 }]
-            }
-        }),
-        egress.clone(),
-    )
-    .await;
-
-    let requests = egress.requests();
-    assert_eq!(requests.len(), 6);
-    assert_eq!(requests[0].method, NetworkMethod::Post);
-    assert_eq!(requests[1].method, NetworkMethod::Patch);
-    assert_eq!(requests[2].method, NetworkMethod::Delete);
-    assert_eq!(requests[3].method, NetworkMethod::Get);
-    assert_eq!(requests[4].method, NetworkMethod::Patch);
-    assert_eq!(requests[5].method, NetworkMethod::Patch);
-    assert_eq!(
-        serde_json::from_slice::<Value>(&requests[0].body).unwrap()["summary"],
-        "Project review"
-    );
-    assert_eq!(
-        serde_json::from_slice::<Value>(&requests[4].body).unwrap()["attendees"][0]["email"],
-        "ada@example.com"
-    );
-    assert_eq!(
-        serde_json::from_slice::<Value>(&requests[5].body).unwrap()["reminders"]["overrides"][0]["minutes"],
-        10
-    );
-}
-
-#[tokio::test]
-async fn gmail_handlers_use_recorded_google_api_shapes() {
-    let scope = scope();
-    let auth = auth_with_google_account(
-        &scope,
-        vec![
-            provider_scope(GOOGLE_GMAIL_READONLY_SCOPE),
-            provider_scope(GOOGLE_GMAIL_SEND_SCOPE),
-            provider_scope(GOOGLE_GMAIL_MODIFY_SCOPE),
-        ],
-    )
-    .await;
-    let egress = Arc::new(RecordingEgress::with_responses(vec![
-        RecordingEgress::json_status(200, fixture("gmail", "messages_list.json")),
-        RecordingEgress::json_status(200, fixture("gmail", "message_get.json")),
-        RecordingEgress::json_status(200, fixture("gmail", "message_sent.json")),
-        RecordingEgress::json_status(200, fixture("gmail", "draft_created.json")),
-        RecordingEgress::json_status(200, fixture("gmail", "message_sent.json")),
-        RecordingEgress::json_status(200, fixture("gmail", "message_trashed.json")),
-    ]));
-
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        GMAIL_LIST_MESSAGES_CAPABILITY_ID,
-        json!({ "query": "is:unread from:ada", "max_results": 25 }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        GMAIL_GET_MESSAGE_CAPABILITY_ID,
-        json!({ "message_id": "msg-001" }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        GMAIL_SEND_MESSAGE_CAPABILITY_ID,
-        json!({ "message": { "raw": "base64url-rfc822" } }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        GMAIL_CREATE_DRAFT_CAPABILITY_ID,
-        json!({ "draft": { "message": { "raw": "base64url-rfc822" } } }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth.clone(),
-        scope.clone(),
-        GMAIL_REPLY_TO_MESSAGE_CAPABILITY_ID,
-        json!({ "message": { "raw": "base64url-rfc822", "threadId": "thr-001" } }),
-        egress.clone(),
-    )
-    .await;
-    dispatch_ok(
-        auth,
-        scope,
-        GMAIL_TRASH_MESSAGE_CAPABILITY_ID,
-        json!({ "message_id": "msg-001" }),
-        egress.clone(),
-    )
-    .await;
-
-    let requests = egress.requests();
-    assert_eq!(requests.len(), 6);
-    assert_eq!(requests[0].method, NetworkMethod::Get);
-    assert!(requests[0].url.contains("/users/me/messages"));
-    assert!(requests[0].url.contains("q=is%3Aunread%20from%3Aada"));
-    assert!(requests[0].url.contains("maxResults=25"));
-    assert!(
-        requests[1]
-            .url
-            .contains("/users/me/messages/msg-001?format=full")
-    );
-    assert!(requests[2].url.ends_with("/users/me/messages/send"));
-    assert!(requests[3].url.ends_with("/users/me/drafts"));
-    assert!(requests[4].url.ends_with("/users/me/messages/send"));
-    assert!(
-        requests[5]
-            .url
-            .ends_with("/users/me/messages/msg-001/trash")
-    );
-    assert_eq!(
-        serde_json::from_slice::<Value>(&requests[2].body).unwrap()["raw"],
-        "base64url-rfc822"
-    );
-    assert_eq!(
-        serde_json::from_slice::<Value>(&requests[3].body).unwrap()["message"]["raw"],
-        "base64url-rfc822"
-    );
 }
